+++ packages/core/src/Agent/AnalyzerAgent/index.ts
@@ -1,7 +1,7 @@
 // Generated by polka.codes
 // This file implements the AnalyzerAgent class for project analysis
 
-import { getAvailableTools } from '../../tool'
+import { getAvailableTools } from '../../getAvailableTools'
 import {
   askFollowupQuestion,
   attemptCompletion,
