+++ packages/core/src/Agent/CodeFixerAgent/index.ts
@@ -3,7 +3,8 @@
  * Generated by polka.codes
  */
 
-import { type ToolResponse, ToolResponseType, getAvailableTools } from '../../tool'
+import { getAvailableTools } from '../../getAvailableTools'
+import { type ToolResponse, ToolResponseType } from '../../tool'
 import { allTools, attemptCompletion } from '../../tools'
 import { AgentBase } from '../AgentBase'
 import type { AgentInfo, SharedAgentOptions } from '../AgentBase'
