+++ packages/core/src/Agent/CoderAgent/index.ts
@@ -1,4 +1,5 @@
-import { type ToolResponse, ToolResponseType, getAvailableTools } from '../../tool'
+import { getAvailableTools } from '../../getAvailableTools'
+import { type ToolResponse, ToolResponseType } from '../../tool'
 import { allTools, attemptCompletion } from '../../tools'
 import { AgentBase } from '../AgentBase'
 import { responsePrompts } from '../prompts'
