+++ packages/core/src/index.ts
@@ -4,3 +4,4 @@ export * from './tools'
 export * from './AiService/ModelInfo'
 export * from './AiTool'
 export * from './tool'
+export * from './getAvailableTools'
